requirements.txt
openai
gradio
yfinance
matplotlib
seaborn
numpy
pandas
plotly
pyyaml
mplfinance
mistralai
fastapi 
uvicorn 
sqlalchemy 
python-jose[cryptography] 
passlib[bcrypt] 
sqlite-utils
bcrypt
